Accumulate test helpers into a single place
 - Few test helpers and macros are scattered around `database/memory.rs` and
`wallet/mod.rs`. These are collected in a single place `testutils/helpers.rs`.

 - the `populate_test_db` macro is changed into a function. Internal logic
 should remain same.

 - A new `run_tests_with_init` macro is added in `testutils.helpers.rs`,
 which can run database tests given an initializer function.
