+++ src/database/memory.rs
@@ -475,126 +475,6 @@ impl ConfigurableDatabase for MemoryDatabase {
     }
 }
 
-#[macro_export]
-#[doc(hidden)]
-/// Artificially insert a tx in the database, as if we had found it with a `sync`. This is a hidden
-/// macro and not a `[cfg(test)]` function so it can be called within the context of doctests which
-/// don't have `test` set.
-macro_rules! populate_test_db {
-    ($db:expr, $tx_meta:expr, $current_height:expr$(,)?) => {{
-        $crate::populate_test_db!($db, $tx_meta, $current_height, (@coinbase false))
-    }};
-    ($db:expr, $tx_meta:expr, $current_height:expr, (@coinbase $is_coinbase:expr)$(,)?) => {{
-        use std::str::FromStr;
-        use $crate::database::SyncTime;
-        use $crate::database::{BatchOperations, Database};
-        let mut db = $db;
-        let tx_meta = $tx_meta;
-        let current_height: Option<u32> = $current_height;
-        let mut input = vec![$crate::bitcoin::TxIn::default()];
-        if !$is_coinbase {
-            input[0].previous_output.vout = 0;
-        }
-        let tx = $crate::bitcoin::Transaction {
-            version: 1,
-            lock_time: 0,
-            input,
-            output: tx_meta
-                .output
-                .iter()
-                .map(|out_meta| $crate::bitcoin::TxOut {
-                    value: out_meta.value,
-                    script_pubkey: $crate::bitcoin::Address::from_str(&out_meta.to_address)
-                        .unwrap()
-                        .script_pubkey(),
-                })
-                .collect(),
-        };
-
-        let txid = tx.txid();
-        // Set Confirmation time only if current height is provided.
-        // panics if `tx_meta.min_confirmation` is Some, and current_height is None.
-        let confirmation_time = tx_meta
-            .min_confirmations
-            .and_then(|v| if v == 0 { None } else { Some(v) })
-            .map(|conf| $crate::BlockTime {
-                height: current_height.expect("Current height is needed for testing transaction with min-confirmation values").checked_sub(conf as u32).unwrap() + 1,
-                timestamp: 0,
-            });
-
-        // Set the database sync_time.
-        // Check if the current_height is less than already known sync height, apply the max
-        // If any of them is None, the other will be applied instead.
-        // If both are None, this will not be set.
-        if let Some(height) = db.get_sync_time().unwrap()
-                                .map(|sync_time| sync_time.block_time.height)
-                                .max(current_height) {
-            let sync_time = SyncTime {
-                block_time: BlockTime {
-                    height,
-                    timestamp: 0
-                }
-            };
-            db.set_sync_time(sync_time).unwrap();
-        }
-
-        let tx_details = $crate::TransactionDetails {
-            transaction: Some(tx.clone()),
-            txid,
-            fee: Some(0),
-            received: 0,
-            sent: 0,
-            confirmation_time,
-        };
-
-        db.set_tx(&tx_details).unwrap();
-        for (vout, out) in tx.output.iter().enumerate() {
-            db.set_utxo(&$crate::LocalUtxo {
-                txout: out.clone(),
-                outpoint: $crate::bitcoin::OutPoint {
-                    txid,
-                    vout: vout as u32,
-                },
-                keychain: $crate::KeychainKind::External,
-                is_spent: false,
-            })
-            .unwrap();
-        }
-
-        txid
-    }};
-}
-
-#[macro_export]
-#[doc(hidden)]
-/// Macro for getting a wallet for use in a doctest
-macro_rules! doctest_wallet {
-    () => {{
-        use $crate::bitcoin::Network;
-        use $crate::database::MemoryDatabase;
-        use $crate::testutils;
-        let descriptor = "wpkh(cVpPVruEDdmutPzisEsYvtST1usBR3ntr8pXSyt6D2YYqXRyPcFW)";
-        let descriptors = testutils!(@descriptors (descriptor) (descriptor));
-
-        let mut db = MemoryDatabase::new();
-        let txid = populate_test_db!(
-            &mut db,
-            testutils! {
-                @tx ( (@external descriptors, 0) => 500_000 ) (@confirmations 1)
-            },
-            Some(100),
-        );
-
-        $crate::Wallet::new(
-            &descriptors.0,
-            descriptors.1.as_ref(),
-            Network::Regtest,
-            db
-        )
-        .unwrap()
-    }}
-}
-
 #[cfg(test)]
 mod test {
     use super::MemoryDatabase;
