+++ src/testutils/mod.rs
@@ -17,6 +17,7 @@ pub mod blockchain_tests;
 #[cfg(test)]
 #[cfg(feature = "test-blockchains")]
 pub mod configurable_blockchain_tests;
+pub mod helpers;
 
 use bitcoin::{Address, Txid};
